app/requirements.txt
# --- Core web stack (matriz estable con Gradio 4.42) ---
fastapi==0.111.0
starlette==0.37.2
pydantic==2.6.4
uvicorn[standard]==0.30.6
gradio==4.42.0
# ---------------------------------------

typer==0.12.3
rich==13.7.1
tqdm==4.66.5

# LangChain + VectorStore + Embeddings
langchain==0.2.14
langchain-core==0.2.36
langchain-community==0.2.12
langchain-text-splitters==0.2.2
faiss-cpu==1.8.0
sentence-transformers==3.0.1

# PDF + OCR
PyMuPDF==1.24.8
pillow==10.4.0
pytesseract==0.3.13

# Audio -> Texto (Whisper local)
openai-whisper==20231117
torch==2.2.2
torchvision==0.17.2

# (Opcional) CLIP
transformers==4.43.3

numpy==1.26.4
